requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
idna==3.10
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-email==0.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.0
